A Rough Skeleton for Popup.html
This is NOT working due to the security reason. Details and solution are here: https://developer.chrome.com/extensions/contentSecurityPolicy#JSExecution
This popup should be able to take in information and send it to database if the security problem is solved. Besides, more forms can be created by simple copy paste later.

Everything besides popup.html and savaData.js contains nothing.